+++ simple_ddl_parser/ddl_parser.py
@@ -69,9 +69,11 @@ class DDLParser(Parser, Dialects):
         return t
 
     def process_body_tokens(self, t: LexToken) -> LexToken:
-        if (self.lexer.last_par == "RP" and not self.lexer.lp_open) or (
-            self.lexer.after_columns and not self.lexer.columns_def
-        ):
+        if (
+            self.lexer.last_par == "RP"
+            and not self.lexer.lp_open
+            and not self.lexer.in_alter_column_definition
+        ) or (self.lexer.after_columns and not self.lexer.columns_def):
             t = self.after_columns_tokens(t)
         elif self.lexer.columns_def:
             t.type = tok.columns_definition.get(t.value.upper(), t.type)
@@ -249,7 +251,7 @@ class DDLParser(Parser, Dialects):
         if t.type in ["RP", "LP"]:
             if t.type == "RP" and self.lexer.lp_open:
                 self.lexer.lp_open -= 1
-                if not self.lexer.lp_open:
+                if not self.lexer.lp_open and not self.lexer.in_alter_column_definition:
                     self.lexer.after_columns = True
             self.lexer.last_par = t.type
 
@@ -258,6 +260,10 @@ class DDLParser(Parser, Dialects):
 
         if t.type == "ALTER":
             self.lexer.is_alter = True
+        elif t.type in ["CHANGE", "MODIFY"] and self.lexer.is_alter:
+            self.lexer.columns_def = True
+            self.lexer.after_columns = False
+            self.lexer.in_alter_column_definition = True
         if t.type == "LIKE":
             self.lexer.is_like = True
         elif t.type in ["TYPE", "DOMAIN", "TABLESPACE"]: